moved main dashboard component into pages directory
